hide irrelevant actions on some pages
